fix: resolve page stuck on 'Umi Loading...' issue
Root cause: app.tsx exported unsupported functions (getToken, setToken, removeToken)
which caused UmiJS plugin registration to fail with error:
'register failed, invalid key getToken from plugin app.tsx'

UmiJS only allows specific convention exports from app.tsx:
- getInitialState
- layout
- request
- etc.

Changes:
- Create src/utils/auth.ts for token management functions
- Update imports in app.tsx, requestErrorConfig.ts, AvatarDropdown.tsx, login/index.tsx
- Remove unsupported exports from app.tsx
