+++ src/requestErrorConfig.ts
@@ -1,8 +1,8 @@
-﻿import type { RequestOptions } from '@@/plugin-request/request';
+import type { RequestOptions } from '@@/plugin-request/request';
 import type { RequestConfig } from '@umijs/max';
 import { history } from '@umijs/max';
 import { message, notification } from 'antd';
-import { getToken, removeToken } from './app';
+import { getToken, removeToken } from '@/utils/auth';
 
 const loginPath = '/user/login';
